Fixed shell crashing (DB#375060)
Replaced the previous "fork and ignore exceptions" with a call to
subrpcess.call. This means that caply no longer supports Python versions
less than 2.4, as subprocess isn't available in them.

For some reason the SIGCHLD was ignored completely, and for some reason
this caused subprocess.call to die. Would need more investigation.