Alsa support from Tom Adams
Might not be very robust yet, has some weird problems with Ubuntu Natty:
if the volume is set to zero, it can't be reenabled via cplay. Note that
this actually happens with alsamixer too. The only way to get sound back
is to use PulseAudio aware controls.

Works in Maverick, though.